backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
python-multipart==0.0.6
aiosqlite==0.19.0
asyncpg==0.29.0
pytest==7.4.4
httpx==0.26.0
gspread==5.12.0
google-auth==2.26.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
